requirements.txt
playwright>=1.46.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.4.0
selectolax>=0.3.23
beautifulsoup4>=4.12.3
tenacity>=8.3.0
python-dateutil>=2.9.0
google-api-python-client>=2.137.0
google-auth>=2.32.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
orjson>=3.10.7
tqdm>=4.66.4
